tests: fix fake_pha test failure on crates
Turns out that crates raises a different error than pyfits, so
write some horrible code to catch the differences. I do have some
I/O work which might avoid the need for this, but that's a long
way from getting merged, and we are always going to have odd
cases like this (without using a very-restrictive I/O layer,
which I don't think we want).