model: allow guess to be used on composite models
The idea is that we just pass on the guess call to the components.
There is no attempt to recognize that this is composite system -
that is, we do not try to go "oh, this is mdl1 + mdl2 so we should
just halve the normalization guesses", because

- there's no way to identify normalization parameters
- there's no a-priori way to determine that a 50/50 split is
  meaningful
- we would really want a way to combine the guesses and use them
  on the composite model, but there is no way to do this with our
  current system (and no obvious way to change our system to do
  it)

We also add guess methods to those model classes where we know
there's no meaningful guess approach.

It is not clear if this approach makes sense. Perhaps we should
do all the checks and reporting in sherpa.ui.utils.guess and
make the non-UI user just call guess on the individual components
by hand (but then how do you send around the data's to_guess
output?).

There is minimal changes needed to get this to work in the UI
layer. I had thought I would need to add something to the
instrument responses - e.g. sherpa.astro.instrument.RSPModel and
the like - but they extend CompositeModel so will hopefully
pick up this logic (although how it interacts with the other
parent, ArithmeticModel, is TBD).