random: better use of overload
As discussed in the review, we don't wnat to type the actual
model, just the previous overload lines. This reduces noise and
the chance of confusion.

I have realised - thanks to pyright complaints - that my types
were not correct, since setting size=None means we get a scalar
out, not an ndarry. So, add overloads to account for this. To
do this means giving a type for the size parameter, which I had
just gone to great lengths to say I wasn't going to type!