parallel: improve typing rules
Switching to pyright suggests some more funkiness with the
Protocol support

- more checks that context is actually defined, which we know
  but the typing code does not

  (really the code should be re-written to make this obvious
   but I'm not sure it's worth it)

- some more funkiness with the Process member of SupportsContext,
  since the "base" version is not present.

The module does not completely pass pyright because I do not
understand what parallel_map_funcs is meant to be doing, and
so have not spent much time addressing typing issues in it.
Note that mypy does not complain because I have not added any
typing rules for the function!