ui: improve guess logic when called with only an id
The ui.guess routine does not match many of the other routines
because insead have having a signature

    (id, some_val=None, ...)

where we swap id and some_val if some_val is None, we have

    (id=Nome, some_val=None, ...)

and so our simple check does not work. So, we know check whether
id is a known dataset identifier before switching. It's not ideal
but should be a good compromise. Before this fix it would not
work.