model: internal change to parameter access
This mainly adds a few comments to the code, since I had thought
I could move the creation of the _par_index directory to the
Model.__init__ routine, but that failed miserably. Hence the
addition of comments.

In doing so I sligthly re-worked the logic for creating the
warning message (when an alias is found) as there's no benefit to
creating a local routine that is only used once.

In doing that I decided to report the user-requested name rather
than the lower-cased version - which is a minor usability change.
This then lead to looking at the setattr routine as well as that
had been setting the name to lower case before the check, but that
is not needed as the getattr routine does it for us. There is
a surprising amount of noise in the tests to accomodate this,
just because the assumptions used to write the test (that the
name would be done in lower case so we don't need to send in the
actual name) is no-longer true.