stats: make calc_staterror a staticmethod by default
To please pyright we can either

- mark all calc_staterror methods with @staticmethod
- remove @staticmethod from those classes that set it

Both approaches are problematic:

- UserStat needs calc_staterror to have access to the object
- if we remove staticmethod a number of tests fail; most
  of these could be adjusted, but not obviously all
  [see below]

So I have gone with "essentially option 1" as it

a) doesn't change the tests
b) reduces the pyright problems with the module to just the
   UserStat calc_staterror method

Test failures when removing staticmethod
----------------------------------------

One set of failures if we remove the use of @staticmethod
support is with Chi2XspecVar, since we currently have to
split the logic between the Stat object and the DataPHA
object to handle background datasets. Really we should
have the logic in the stats class, but this would be
very invasive (and would need to work out how to handle
scaling between different background datasets). At the
moment the DataPHA.get_staterror method has a special
case for the staterr function matching

    Chi2XspecVar.calc_staterror

but this only makes sense when it's a staticmethod. As
soon as it becomes a general class method the handling
of Chi2XspecVar will fail and I just do not want to
go and try to address that failure here.

Note that the above is related to issue #356