tests: add regression tests for multinormal/t_pdf
It is not clear if we should just remove these routines, but
let's at least add some basic regression tests as changes in
NumPy are going to cause changes to the code (e.g. drop of
matrix support).

In fact, if we had added these tests as part of #2025 we would
not have made the "thinko" that caused the multit_pdf error to
creep in in changeset c3700065f4f06e3b9cb3eb7e669fea1bea3d5dae