utils: address issue 2025
Avoid the use of NumPy matrixes for the multit/normal_pdf
routines. Fortunately the change is minimal - just be explicit about
where matrix multiplication takes place by signalling it with @.

This also avoids a NumPy 1.25 warning message, since the previous code
was calling

    float(array_of_size_1_by_1)

which now causes a warning message, and could be addressed by saying

    float(array_of_size_1_by_1[0, 0])

but we now create a scalar value rather than a [1, 1] array, so
there's no need to change this part of the code. We do not remove the
associated warning in sherpa/conftest.py since there are other places
in the Sherpa codebase where we still cause the warning to be created.