utils: add annotation support to export_method
This is to fix #2053. I have tried with Python 3.9 and that
seems to support __annotations__, so I have no tried to make
this "optional" (i.e. handle a TypeError when the field is
accessed).

I have added some documentation to mention some exploration I
tried to replace the existing code with functools.wraps, since
this would presumably make the code more future proof (since
we just rely on Python itself rather than having to do our own
deconstruction and reconstruction of the code), and fix issues
like #1901. Unfortunately the functools route does not address
the reason this routine was created - to hide the fact this is
a method when raising a usage-error message.