guess: adjust XSPEC normalization guess
Change how the norm parameter for XSPEC additive models is guessed.
We now evaluate the model (assuming the dependent axis units are
sensible which they will not be when using channel units) and then
calculate the "guessed" normalization by summing up the model and
comparing it to the data. This is an alternative to using the maximum
level, and it is not clear what the best approach is.

The old norm guess didn't seem to work well on the one test I tried
when developing this approach.

See issue #104 which used a different approach (not obviously
possible here with the way guess is set up).