model: add basic typing to the regrid module
In adding typing rules it's become obvious that some of the
methods are not clear on the behaviour when the axis is empty.
For instance, should grid error out or return None (or a tuple
of None values)?

So try to make the code a bit clearer as we add types.