model: change typing rule for model evaluation
Use the overload decorator from typing to point out that when
calling a model we have

 - no arguments -> returns itself
 - arguments    -> returns the model evaluated on the grid
   (this gets messy because we still have the "input/output
   can be a sequence or ndarray" issue that complicates the
   typing)

Previously the typing rule didn't capture that nuance. This
becomes important when evaluating calls to models; whether it
actually helps (i.e. let's us igonre the "returns self" condition
without some help) is to be seen.

As part of this we now mark typing_extensions a runtime
installation requirement. This is required by many packages
and so I do not think it really adds much (i.e. a user will
probably already have this). We could try to work around this
but I've hit this issue a few times now and it seems this is
the most-sensible (and aproved) approach.

This has shown some more thinking is needed for the NestedModel
class - which I think is currently only used to apply the JDPileup
model. Issues include

- this shares some functionality with convolution-style models,
  should there be more shared code

- how do we best type models like JDPileup and the convolution
  models, since they take an extra argument

There is also some thinking needed to decide how best to type
the ModelDomainRegridder<n>D class, since we want to avoid
circular imports but the use of a Protocol has lead to more
warnings from the typing code.