fit: ensure fh is closed on exit
We can't make use of a simple context manager for handling
the fit output since

a) we may not have a handle/filename
b) even if we do, we may not want to close it (i.e. if sent in by
   the user)

We can at least try to emulate the "close even on an error" handling
of a context manager with try/finally, as suggested by @hamogu

This is a bit invasive visually, but it's not clear to me that
breaking the code up into smaller chunks is worth it here.