fit: ensure the statistic object is consistent (issue #2063)
We can make sure that changing the Fit stat also changes the
associated _iterfit.stat field, which seems like it should be
the correct thing to do.

There are only two tests where this is an issue, and it's unclear
whether the changes are good, bad, or not significant, since

a) the tests appear to be regression tests
b) we don't know how the test data was created, so we don't know
   what the true values are

   Note that the parameter values do not appear to have changed
   "hugely", but I haven't looked into the results too deeply.

The tests do show the change - in that the statistic value reported
in the "fit(outfile=...)" option now remains consistent (i.e.
matches the expected value, assuming our new interpretation of
the stat field of the Fit object is correct [which it should be,
I claim]).