TESTS: add a test of issue #2063
The issue at hand is that if you *change* the stat method of the
Fit structure then you will get a mis-match with the actual
fitted statistic, which remains using the _iterfit.stat method
(i.e. the original statistic).

We only have two test cases that expose this difference, so take
advantage of the new "write the fit results to StingIO" capability
and show that the stat value has changed (i.e. all but the last row
use the original statistic, in this case Chi2DataVar, so the stat
value is ~ 8500, and only the last row uses the Cash or CStat
value).

This is intended as a regression test.