IO: minor clean up
I have found that numpy 1.24 changed the asarray behavior: prevously
it would create a "ragged" array (i.e. allow each row to have a
different number of columns) but it would complain about it. From 1.24
it now errors out (unless you force the output type to be
object). There are several places where we have code that may or may
not assume we have a n by m grid of data, and the code to do
so (e.g. whether we need to be explicit and check or just let numpy
error out) therefore depends on the numpy version in use. So the code
attempts to be explicit (i.e. check itself) first, so we can use pre
and post numpy 1.24 (although numpy 1.23 is soon going to be marked
end-of-life).