model: internal very-minor cleanup
There is no change in behaviour here, but the new code is
slightly better, since we are actually checking the thing
we care about (if tp.link is not None) which was previously
an indirect check (since tp would only be set to 'linked' when
to was not a link).

This was found by the typing checks, but pulled out into a
separate commit.