root: allow root-finding code to work with NumPy 2.0
As noted, NumPy has changed how type conversion works - e.g. if
a value includes a np.float32 or np.float64 value, what is the
result of an expression that uses that value. It turns out that
the zeroin method relied on the pre-NumPy 2.0 behaviour, and
we can retain that with NumPy 2.0 by explicitly converting the
np.float32 value (DBL_EPSILON, which is set to the .eps value)
to a float.

I have spent a lot of time trying to track this down, and I am
happy to take this fix, as anything more would just not get done
in time for the next Sherpa release, given the lack of commentary
on this code.