fit: ensure parameter values are stored as floats
In NumPy 2.0, float64/32 values are displayed as 'np.floatxx(...)'
rather than '...'. This causes the __str__ output of a number of
our routines to change. It shouldn't affect the behaviour of
code, but there could be code that is parsing the string output
for some reason, and it is visually distracting, so ensure the
parameter values, mins, and maximums, are cnoverted to float.

However, in doing this I have found that the error estimation
code does not always seem to set the error flag, which is used
by ErrorEstResults to store the results. The problem was that
the new

    <list>.extend(float(<value>))

calls would fail when <value> is None, which one of our test
cases hits. So there's three things:

1) what to do in ErrorEstResults

a) I could have added an extra check to see if the value is
   None to avoid the float call, but not record this as a
   warning

b) or just explictly check for it in the current code, and so
   generate a warning which would not have been created before

I have gone for option b as it seems the safer route.

2) Why does the error flag not get set?

This appears to be in the confidence routine. Perhaps we should
remove this logic from the error estimation routines and just
have it in the relevant Fit class? I have spent no thought on
this.

3) Why do we only have one test that seems to cause this?

It is actually in a test which we expect to "fail" in some
manner, and really the code should never get as far as trying
to calculate errors, but that is a different issue (already
documented in the code).