optfcts: remove order=C/F when unnescessary
a) the array is meant to be 1D so it doesn't matter if the
   ordering is FORTRAN or C

b) all values are the same (i.e. np.ones() * value) so it shouldn't
   matter

Rather than use np.ones() * val, replace with np.full(). The
speed advantage should be irrelevant here, but let's try to
write code we can copy elsewhere.

There is one other place where "order=F" is used in this module but
I assume that is intended, unlike these cases.