plot: rework the plot structure WIP
Separate the low level classes, that actually do the plotting,
from the user-callable classes that are sent the data objects.

  - this is hugely invasive
  - since the "classes that have a prepare method" all end up calling
    these base classes, we know have them stored in the _object field,
    primarily so that the plot/contour calls make sense (as they have
    very different signatures)
  - when doing this I am re-working some classes that have previously
    used multiple inheritance when it's not obvious they really take
    advantage of it. Some examples are: FitPlot, SplitPlot, and
    Confidence2D.

Note that some classes no-longer derive from NoNewAttributesAfterInit
and have not been set up to use slots to restrict attribute access.

The new approach means that we need to be more careful about
merging the stored preferences with those from the user. I think
there's at least one astro case where the old code wasn't doing this,
which might have been a bug. The new code is a bit annoying and
repetitive.

Is this sensible?

What is missing?

Should we export the old names for existing code?

Can we remove some code (e.g. the preferences/settings for the
Object classes, as well as their overxxx() methods)?