modernize test_model.py
This replaces numpy with np and uses parameterization to reduce the amound of code in tests.
I do this because in the next step I want to copy and paste on of those tests for a different situation, so I want to clean it up first to make the amount of code I need to copy more managable.