Calculate absorption models from just one value
This introduces a new type of cache, that, instead if caching the model value for each parameter value, cached is for the parameter value=0.1. While the cache in general does not know about which model it is applied to, this is specifically designed for absorption models in XSPEC, which have the form
out(E) = exp(-N_h * sigma(E))
Those models are relatively expensive to calculate because it requires evaluation of a large number of atomic cross-sections. So, instead, we just evaluate them once.
The fiducial value of 0.1 is physically motivated, because a larger number (say N_h=1) already gives no throughput at low energies, and we end up with values of 0, which causes problems for log(0).

This new model is NOT applied to "XSwabs", which is an obsolete model used for historical purposes only and seems to be to inaccurate internally, that the absorption calculated with that law, is not actually an exponentia. It should be, according to the XSPEC docs, but numerically it's not. Instead of debugging that and making choices on how to deal with it, it's simply not touched in this PR.