plot: abstract out the display code
The string representation follows pretty-much the same form for
all the classes, so move to a more-standardised representation,
where the ordering is set by the _fields field and then a
helper routine (display_fields) that does the actual string
creation (and that can then be used in the various __str__
calls). This is a "minimal" change, as we could instead try
to add this logic as an ABC or Protocol, but I mam wary of
making a large change to the classes (as we have a bunch of
other changes we could make to the class structure, so I
would rather wait until we decide to do that). For now I
prefer the simpler approach taken here.

This is similar to the way the Data code handles the string
representation, but we do not share code as it's not exactly the
same and it doesn't seem worth combining the two.

One "issue" is that I use a simple Hungarian Notation style
for the elements of the _fields array: if they end with a '!'
character then the field is displayed as a scalar, otherwise
it is assumed to be a 1D NumPy array. This was the simplest
scheme I could come up with and doesn't need to need more
additions (at least with the current set up).

I note that several tests have had to be updated now we iuse
a more systematic style (it is just addding more spaces between
the field name and the '= ...value...' section).