plot: make more use of super
With the recent changes to the plot hierarchy it's easier than
it used to be to change to using super().foo to call the foo
method of the superclass. So take advantage of this. As part
of this

a) can call super().plot(overplot=...) rather than
Plot.plot(x=..., overplot=...)

b) it is more obvious when we are intentialy breaking the
superclass chain and calling the base object (which takes
the data to plot) rather than the the parent class (which
doesn't take these arguments).