Test: add a check for FitPlot
The behaviour of JointPlot depends on what it is sent for the
top plot, when determining how to hide the x-axis label. The
check tries to be faithful to Python's roots, and does not use
an isinstance check (unlike much of Sherpa!) to determine whether
we have been sent a FitPlot or not. This check, as currently
written, requires that FitPlot does not have a xlabel attribute,
so we add an explicit test for this (and add in a few other
checks as well).

This is a regression test, as

- there is no real reason why FitPlot can / can not have this set
  of attributes

- the JointPlot code could be written differently, or we could
  add a set_<label> routine to the plot classes so that FitPlot
  can pass it to its contained classes

I have spent a day trying to track down changes to the xlabel
value of joint plots that has ended up being because of this
issue, so I want a test / check so we know if it gets changed
again.