IO: clean up irregular-grid error reporting
In numpy 1.24 it became an error to pass in irregularly-gridded
data to asarray. Prior to that it would return an ndarray with a
dtype of object (and generate a deprecation warning; and presumably
in even older versions something different happened).

To avoid having to deal with the various combinations, we check
the column sizes ourself (for 2D grids only) and throw an IOErr
so it should be recognizable.

There are also some changes made to try and appease the typing
gods, as I haven't yet worked out how best to handle the
"can be a sequence or ndarray" type.

A few tests need to be changed since the error type and/or message
has changed, but these are regression tests, so as long as we
still error out then this is good.