IO: treat a HDUList as a context manager (pyfits)
In #2061 I introduced the _get_file_contents helper routine which
returned both a fits.HDUList and a boolean, where the latter decided
whether the former should be closed or not. During a review @hamogu
asked if this could be done with a context manager, but I could not
see how.

Well, now I see how: with the nullcontext handler (which just makes
something act like a context-handler but not actually doing anything),
coupled with the fact that fits.HDUList appears to act as a CM
anyway. This means we can change

    hdus, filename, close = _get_file_contents(arg)
    try:
        ...
    finally:
        if close:
	    hdus.close()

with

    cm, filename = _get_file_contents(arg)
    with cm as hdus:
        ...