data: allow axis labels to be changed
To complement

  get_xlabel, get_x0label, get_x1label, get_ylabel

add

  set_xlabel, set_x0label, set_x1label, set_ylabel

for the data types. This is trickier than at first thought, since

a) we introduce _xlabel, _ylabel, ... fields to store the labels
   so we can set up the defaults (in the __init__ call) and then
   over-ride them by the user

b) since we have some interesting hierarchies, with __init__ calls
   that can not be just passed to a superclass, we have to either
   create these attributes in more places than we'd want

c) some classes have labels that depend on the object setting
   (DataPHA and DataIMG), so we need to decide how to handle this.
   The assumption taken is that if a user has not called set_*label
   then the get_*label call returns the expected value, but as soon
   as set_xlabel is used then the requested label is always returned.

I thought about adding in UI routines to provide easy access to the
user, but for now we rely on the user manually having to call the
get_data routine for us.