fit: simplify setting of file handle
There is no functional change here. The code has been changed to
use a context manager to handle closing the file handle during
a fit. This wasn't done originally because we do not always want
to close the handle - e.g. if the user sends a file handle or
sends in None - and so I had come up with code that works but
is ugly, as noted in reviews by @hamogu in an unrelated PR that
had similar behaviour. I realised, in #2017, that we can use
the contextlib.nullcontext handler to handle the "do not close"
case, so apply that logic here.

Ome issue is the typing rules, as it's not clear what the best
approach is for typing a context handler, so we punt on that.