data: better support for filtered asymmetric errors
This is a **minimal** change to allow plotting filtered data
with asymmetric errors.

There are a number of methods in Data1DAsymmetrucErrs that
should also be looked at, but as this code is just not well
documented or tested I am going for minimal changes.

I note that some of the plot backends fail here: one is a
testing backend, and one is bokeh.