docs: add a notebook showing the asymmetric errors
This is a **very** minimal notebook, and basically shows the
absolute bare-minimum. However, I don't really understand what
it's meant to be doing so I don't have much to add. There are
also a lot of questions that writing this notebook raises,
such as what the fit statistic/method actually is and why can
it not be changed, that I do not have the time or energy to
look into.

This notebook could also go through the "do this with the object
interface", but I don't have time to add that here.