io: better handling of warning / error messages from I/O operations
Recent changes have tried to

- provide more context in an error message (e.g. try to say what
  file is being looked at, since the message from the backend may
  not tell us

- provide a more consistent behaviour between the two backends.

Looking at the results of #2017 from some CIAO regression tests,
I have come to the conclusion that

- if possible, we add the context in the "higher level code" and
  don't add too much context in the backend, because the backend
  code is used in multiple cases, and so does not know the context

- at some point we have to understand that the backends are
  different

So in this commit, when we turn an error message into a warning
we can add context - i.e. what file type (ARF, RMF, background)
has the error. This loses guaranteed knowledge of the file name,
but we actually didn't have this here anyway, since you get
different messages if you try to load a ARF as a RMF, for instance.

We also clean up the xstable model error to provide context.

This is unfortunately not a top-to-bottom revamp of the I/O
code, as I don't have time to do this as other parts of the I/O
code need work more urgently than "improving error messages". It
is also something that needs thinking about at each use site, and
not - I **believe** - done in a generic manner.