data: minor clean up of RMF filtering
Add types to sherpa.astro.data._notice_resp which lead to types
being added to sherpa.utils.filter_bins. A possible error
condition is now recognized (and you get a slightly useful
error message, rather than the previous NumPy error case, but
although what exactly should happen here is unclear as we do
not have good tests or documentation on this area of code).

There is a small change to the code to make it a bit-more
obcious the behavior in the "has only one of ARF or RMF" case.
I assume this is completely untested for the ARF-only case.

I note that the error conditon was noted by pyright as I hadn't
seen it.