data: fix some typing rules
The get_xerr type was typed confusingly - did it return
Optiona[np.ndarray], np.ndarray, or None. It is really the first
option, so fix those.

We add types to get_xlabel and get_ylabel and change the get_ylabel
calls to match the superclass parent - that is, we allow a
yfunc=None argument that is unused and should probably be removed.