model: avoid a compilation warning
There can be a complaint when compiling that the
py_integrated_1d symbol is unused. I assume this is actually an
issue with the compiler - in this case g++ - since it is used,
but maybe the fact it's used within a template leads to a false
positive.

A simple solution is to just inline the function, since - to be
honest - moving it into a separate routine does not make a huge
difference to the reader of the code.