data: fix up label typing rules
For the set_ylabel calls make sure we use the same argument
as the parent classs, which has an optional yfunc argument
that is unused. I really should just remove it but it does
not seem worth it.

I'm not sure what happened here for the x*label stuff, but
fix the typing.