XSPEC: XSPEC models now have a spectrum argument
This attribute is sent to the model function as the spectrumNumber
argument, which allows users to change the value from the value
of 1.

Since the model evaluation cache is going to be invalidated when
the spectrum Number is changed, we clear it whenever the value
is changed. An alternative approach would be to include the spectrum
number in the cache calculation, but that would be a significantly
larger change. This should work for most use cases (at least until
we better understand how these models are used).

There is some surprising test differences here, in part because
we have not managed to fix the XSPEC regrid issue and the addition
of the spectrumNumber argument means that the error that is thrown
in these regression tests needs to be updated.

There are also some explicit changes needed to the spectrumNumber
handling since we now take the value from the spectrum attribute
(**IFF** not set), which changes some of the tests (as to what
spectrumNumber is actually used).

One issue is whether we want to pass the spectrumNunber value
through "complex" models - such as convolution or binary op
models. My current thinking is NO, because we want the value to
be taken from the model evaluation where possible. This lets the
user say "this component needs to be associated with this spectrum
number", and we can avoid passing this through the system. One
advantage is then you have less worry about what to do if there's
a mismatch - e.g. mdl1 + mdl2 where they both have different
spectrum numbers; with this approach we just let them be but if
we wanted to pass on a "spectrum number" to the composite model
then what woould be used?