ASTRO: initial support for the XFLT keyword routines
This provides the first version of an API for the user to use
the XFLT keywords. The idea is that the interface works on the
idea that we care about sending in all the settings at once,
and then we have a routine which does this when given a PHA
header.

As noted, it might make sense to make the spectrum identifier
a separate type, so separate it from the general IdType we are
using, but let's leave that for a later revision after we have
some experience with this.

There may need to be some work to change reporting, such as

  - from the load_xsxflt routine,
  - and add some way to display the existing settings

as well as tweaks to how we parse the XFLTxxxx values, as
XSPEC appears to do things in an "interesting" manner for
strangely-formatted values (is it an error or do we just
replace it with 0? The documentation is non-existant here).