XSPEC: build smaug, polconst/lin/pow models
Now we technically support models which need to be re-evaluated
per spectrum (via setting XFLT keywords), provide low-level
access to the four models (oe additive, three multiplicative).

Note that these have not been tested *AND* the polarimetry
models are likely not to be useful as we need other changes
to support using these models (since they need to be used
with multiple datasets and the errors do not add as they would
with normal data).