XSPEC: add show_xsabund command
This adds

 - sherpa.astro.ui.show_xsabund
   (and also to sherpa.astro.ui.utils.Session)

 - sherpa.astro.xspec.get_xsabundances()
   and, not exported, sherpa.astro.xspec.get_xselements()

The aim is to provide a "useful" show_xsabund() command for the
UI user, but the get_xsabundances() command may be useful too.

Unfortunately, with the current interface, we can't get at the
description/documenatation for each abundance table, so we can
only label it as "angr" and not "angr: Anders E. & Grevesse N. ...".
Thiscould be done but would require addung functionality to the
compiled code in sherpa.astro.xspec._xspec.

This code is based on the larger "take advantage of the
FunctionUtility interface" PR #1615, which is where a number of
the tests are taken, that cover some functionality not actually
changed here.