ui: resample_data now uses the chosen stat/method
We now use the chosen stat and optmizer in the simulations
done by resample_data. Previously they were hard-coded to
least squares statistic and LevMar optimizer.

We could have made the user change them with arguments to
this call, but it does not seem worth it, and doesn't really
fit in with other parts of the UI (although it would be easy
to make them optional so they can be easily over-ridden if
wanted, but is it worth it?).