ui: reset the RNG state in a better place
Functionally there's no change in the code behaviour, but it is
more obvious here, as the RNG setting has nothing to do with the
image types. I assume that I added the set_rng call in this
location either due to a botched rebase or I just got confused;
my money is on the former as the image/plot code saw a lot of
work to get it ready for merging.

This also adds some typing rules to the set/get_rng routines.