XSPEC: add set_xsabundances
This follows get_xsabundances, added in #2107, which allows the
user to set the abundances from a "vector" (in this case via a
dictionary). This was taken from #1615, where the implementation
is nicer - since we can use the FunctionUtility interface - but
we can provide a routine with the same API even if internally it
is a but messier.