IO: ensure NUMELT/GRP are written out as integers for RMF
For some reason the NUMELT keyword was being treated as a
floating-poiint value when it should/must be an integer, so
fix this. Unfortunately this is tricky to identify in a test
as we don't explicitly check the on-disk formats, as we are
much-more concerned about being able to read back in the
files we have written out, and in this case float<->int conversions
can just happen automagically.