IO: do not try to change the CHANNEL column of PHA files
When reading in a PHA file, do not renumber the CHANNEL column
if it starts at 0. This makes round-tripping data less
confusing, and will lead to better handling of such data
once the RMF code is updated.

There is one user-visible change, since notice/ignore filters
given in channel space will need to be updated (subtract 1
from the upper and lower edges) when reading in such data.