scripts: tweak check_xspec_update.py
This script now uses argparse so it is a bit more usable (although
this is only meant for a very-specific use case).

Added the --noswitch flag to make it easier to skip the switch
parameters which - in the model.dat file - have no limits but we
have actually given a range (based on the documentation).

Co-authored-by: wmclaugh <wmclaugh@cfa.harvard.edu>