XSPEC: add 12.14.1 model: sssed
The sssed model appears to have a messed up frac_th line in the
model.dat. I have reported this to the XSPEC team and have guessed
at what the values should be - the issue is that it looks like they
used the "query for parameter in xspec" line of

    delta then limits....

but the model.dat file takes

    limits... then delta

I have had confirmation frmo the XSPEC team that this is indeed
wrong, and it has been fixed in patch a for XSPEC 12.14.1.