XSPEC: add 12.14.1 models: vagauss/zvagauss
This pair is interesting since they (for the default parameter
values) only produce emission over a very-small energy range
(1.235 to 1.245 keV or so), and this happens to match one of
the excluded energy ranges I used in one of the tests. The
easy solution is to just skip these two models for this test,
as I don't want to go to the effort of re-working them only to
have it cause issues in other models.