XSPEC: add 12.14.1 models: vgaussian, zvgaussian
So, XSPEC labels these vgauss, zvgauss on the HTML pages but
vgaussian, zvgaussian in the model.dat file.

I think we have something similar for the existing gaussian and
zgaussian models, but in this case we have used XSgaussian
and XSzgauss as the model names. It appears. I do not have the
energy to track it down.