docs: update fake_pha examples
The fake_pha example code does not match recent behavior, so
try to fix it. Unfortunately it's a bit hard to since we do
not currently generate the plots/output "on the fly", and so
some examples are hard to re-generate. Therefore the examples
have been re-written to try and capture the correct behavior,
but they need a review.

Ideally we would build these plots when creating the documentation,
but it turns out to be hard to do, since I want to read in example
data files from sherpa-test-data, and that really complicates
the setup. So we do not do that here.