build: hide ipynb notebooks copied over to the docs directory
The system used to copy over the noteboks to the docs directory
is a bit clunky. One part of the clunkiness is remembering to
update .gitignore whenever we add a new notebook. Well, we can
just use a blanket ban to avoid this, at the expense of having
to remember that any notebook we may add directly has to be in
a sub-directory of docs/.