data: add a get_backgrounds method
This provides a more-useful API than the existing

    for bkg_id in foo.backround_ids:
        bkg = foo.get_background(bkg_id)
        # we know bkg is not None but the types do not ...

There are still places where background_ids is useful, and we
do not want to break existing code. There are some isses that
we *could* look at into how background_ids is created/stored/...
but I do not think it's worth it at this time.