Fix #2117
We had inconsistent handling of the STAT_ERR and SYS_ERR values,
both recently (changes in PR #1921 to move the deconstruction of
the PHA data from the backend into the generic layer), and also
in the original, per-backend code.

We have

- values can be set as a column or a header value (a scalar)
- the use of 0 to indicate "we don't have a value calculated"

Now we check

- first for the column
- then the keyword

and then check if all values are 0, and if so we remove it
(i.e. do not send it to DataPHA).

This means that the code now behaves slightly differently,
since you may not get a message about "errors being present
but unused" when the error values are all 0, and the column
will no-longer be created when it is effectively unset. This
should **NOT** lead to differences when fitting the data.