UI: add calc_model/source commands
If you want to evaluate a model to match the data then the
easiest way has been to call get_model_plot and then use the
y attribute of the plot object. However, this is annoying
because users do not always want to use the plt options - e.g.
normalize by the exposure time or divide by the bin width.

So, we add

    calc_model()
    calc_source()

routines which evaluate the model on the relevant grid and return

- the independent axis (if integerated will get both edges)
- the evaluated model value

The return values depend in part on the data object - e.g. the
analysis setting for PHA data - and in part the model (i.e. what
are its units).

There is an asymmetry in these two routines, since

- calc_model will filter (and group) the data
- calc_source does not filter (and can not group)

We could apply a filter to the calc_source data to "match" the
chosen data filter, but it is not ideal (since notice/ignore act
on the EBOUNDS energy grid and this is the MATRIX energy grid,
which is technically different).

An alternative would be to follow plot_source and add lo/hi
arguments so a user can say "only give me the data between these
two limits" - but it's easy for a user to do this themselves,
and it wouldn't reduce the time taken to make the call.